requirements.txt
tensorflow-gpu==2.0.0
numpy 
pandas
matplotlib
seaborn
PyYAML
zipfile